% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-checklists-package.r
\name{get_checklists}
\alias{get_checklists}
\title{Retrieve checklists function}
\usage{
get_checklists(start_date, end_date)
}
\arguments{
\item{start_date}{No default. Format: "yyyy-mm-dd"}

\item{end_date}{No default. Format: "yyyy-mm-dd"}
}
\description{
This function allows you to retrieve checklists for any given date range
}
\keyword{FPT}
